requirements.txt
matplotlib==3.7.1
numpy==1.22.4
onnxruntime==1.14.1
pandas==2.0.1
pygame==2.3.0
pyomyo==0.0.5
PyWavelets==1.4.1
pyzmq==25.0.2
scipy==1.7.3
tqdm==4.65.0
